seperating character tool out
