--- com.unity.perception/Runtime/Validation/AIContentData.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5a82ed528e1016448a8fc750ba7a7b62
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
