--- com.unity.perception/Tests/Editor/EditorValidationTests.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f5aa501b7f8ea7643a9e87db785ac439
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
