--- com.unity.perception/Tests/Runtime/Validation.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d83e5825e9a96484aa0918f2b6f69247
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
