--- com.unity.perception/Tests/Runtime/Validation/RuntimeValidationTests.cs
@@ -1,13 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine.Perception.Content;
-using UnityEngine.TestTools;
-using NUnit.Framework;
-
-namespace UnityEngine.Perception.ContentTests
-{
-    public class RuntimeValidationTests : ContentTestBaseSetup
-    {
-
-    }
-}
