--- com.unity.perception/Tests/Runtime/Validation/RuntimeValidationTests.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9a64fa29765ea6147a1804421f556d17
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
