--- com.unity.perception/Tests/Runtime/Validation/TestBase.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7b53ac303673fa243b11391f8528de10
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
